Fix JavaScript key events and tabbing
Allowing more keys to be translated by the AHK installer's accelerator
translation code makes JavaScript's keydown and keyup events to work in
more situations. This commit reimplements the accelerator in a flexible
framework that allows a developer to modify which key combinations are
or are not passed through to the IES control. Additionally, it fixes the
cyclic tabbing behavior that was broken by neutron's original
accelerator translation code.